requirements.txt
streamlit==1.32.0
numpy==1.26.4
pandas==2.2.2
xgboost==2.0.3
scikit-learn==1.4.2
rdkit-pypi==2022.9.5
joblib==1.3.2
